move all methods from buildtest/tools/docs.py into buildtest/tools/tutorialexamples.py refactor codebase by merging implementation of spack and aws tutorial build into one method 'generate_tutorial_examples'. buildtest will only delete example directory when dryrun is not enabled
