+++ buildtest/main.py
@@ -55,10 +55,7 @@ from buildtest.log import init_logfile
 from buildtest.system import BuildTestSystem
 from buildtest.tools.editor import set_editor
 from buildtest.tools.stylecheck import run_style_checks
-from buildtest.tools.tutorialexamples import (
-    generate_aws_examples,
-    generate_tutorial_examples,
-)
+from buildtest.tools.tutorialexamples import generate_tutorial_examples
 from buildtest.tools.unittests import run_unit_tests
 from buildtest.utils.file import (
     create_dir,
@@ -384,10 +381,7 @@ def main():
         )
 
     elif args.subcommands == "tutorial-examples":
-        if args.examples == "spack":
-            generate_tutorial_examples(dryrun=args.dryrun)
-        elif args.examples == "aws":
-            generate_aws_examples(dryrun=args.dryrun)
+        generate_tutorial_examples(examples=args.examples, dryrun=args.dryrun)
 
     elif args.subcommands in ["stylecheck", "style"]:
         run_style_checks(
